dom/media/test/test_bug1242594.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1242594
-->
<head>
  <meta charset='utf-8'>
  <title>Bug 1242594 - Unbind a video element with HTMLTrackElement
  should not remove the TextTrack</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

var video = document.createElement("video");
video.src = "seek.webm";
video.preload = "auto";

var trackElement = document.createElement("track");
trackElement.src = "basic.vtt";
trackElement.kind = "subtitles";

document.getElementById("content").appendChild(video);
video.appendChild(trackElement);

video.addEventListener("loadedmetadata", function run_tests() {
  is(video.textTracks.length, 1, "Video should have one TextTrack.");
  var parent = video.parentNode;
  parent.removeChild(video);
  is(video.textTracks.length, 1, "After unbind the video element, should have one TextTrack.");
  parent.appendChild(video);
  is(video.textTracks.length, 1, "After bind the video element, should have one TextTrack.");
  SimpleTest.finish();
});
</script>
</pre>
</body>
</html>
